+++ examples/vue/src/App.vue
@@ -1,5 +1,6 @@
 <template>
-  <div>
+  <div class="app">
+    <h1>Example Using SDK on Vue JS</h1>
     <table style="">
       <thead>
         <th>Id</th>
@@ -9,60 +10,94 @@
         <th>Version</th>
       </thead>
       <tbody>
-        <tr v-for="block in blocks" :key="block.id">
-          <td>{{block.id}}</td>
-          <td>{{block.previousblockhash}}</td>
-          <td>{{block.height}}</td>
-          <td>{{block.timestamp}}</td>
-          <td>{{block.version}}</td>
+        <tr :key="block.block.id" v-for="block in blocks">
+          <td @click="onClickBlockId(block.block.id)" class="click">{{ block.block.id }}</td>
+          <td>{{ block.block.previousblockhash }}</td>
+          <td>{{ block.block.height }}</td>
+          <td>{{ block.block.timestamp }}</td>
+          <td>{{ block.block.version }}</td>
         </tr>
       </tbody>
     </table>
+    <br />
+    <template v-if="!!detail">
+      <h2>Detail Block</h2>
+      <br />
+      {{ detail }}
+    </template>
   </div>
 </template>
 
 <script>
-import zoobc from 'zoobc'
-
+import zoobc from 'zoobc-sdk';
 export default {
   name: 'app',
-  data () {
+  data() {
     return {
       blocks: [],
-      error: ''
-    }
+      error: '',
+      detail: null,
+      hosts: [
+        { host: '//85.90.246.90:8002', name: 'Demo Testnet 1' },
+        { host: '/n1.alpha.proofofparticipation.network:8080', name: 'Alpha Testnet' },
+        { host: '//45.79.39.58:8002', name: 'Demo Testnet 2' },
+      ],
+    };
   },
-
-  created() {
-    this.listBlocks()
-  },
-
   methods: {
     listBlocks() {
-      zoobc.connection('http://18.139.3.139:7001')
-      zoobc.getBlocks(0, 5, 1)
+      zoobc.Block.getBlocks(0, 5, 1)
         .then(res => {
-          this.blocks = res.blocksList
+          this.blocks = res.blocksList;
         })
         .catch(err => {
-          this.error = err
+          this.error = err;
+        });
+    },
+    onClickBlockId(id) {
+      zoobc.Block.getBlockById(id)
+        .then(res => {
+          this.detail = res;
         })
-      }
-  }
-}
-
+        .catch(err => {
+          this.error = err;
+        });
+    },
+  },
+  mounted() {
+    zoobc.Network.list(this.hosts);
+    zoobc.Network.set(0);
+    this.listBlocks();
+  },
+};
 </script>
 
 <style scoped>
-  table {
-    font-family: arial, sans-serif;
-    border-collapse: collapse;
-    width: 100%;
-  }
+.app {
+  font-family: arial, sans-serif;
+}
 
-  th, td {
-    border: 1px solid #000;
-    text-align: left;
-    padding: 8px;
-  }
+.app h1 {
+  font-weight: bold;
+  text-align: center;
+}
+
+table {
+  font-family: arial, sans-serif;
+  border-collapse: collapse;
+  width: 100%;
+}
+
+th,
+td {
+  border: 1px solid #000;
+  text-align: left;
+  padding: 8px;
+}
+.click {
+  cursor: pointer;
+}
+.click:hover {
+  opacity: 0.5;
+}
 </style>