+++ tests/Network.test.ts
@@ -5,24 +5,33 @@ import zoobc from '../src';
 
 const hosts = [
   { host: 'https://n0.alpha.proofofparticipation.network:8443', name: 'Alpha Testnet' },
-  { host: '172.104.47.168:7000', name: '168 Testnet' },
+  { host: 'http://172.104.47.168:7000', name: '168 Testnet' },
+  { host: 'https://explorer-nodes.alpha.proofofparticipation.network', name: 'Alpha' },
 ];
 
 describe('Network Unit Testing :', () => {
-  it('Set Network should return array of object', async () => {
+  it('should return array of object', () => {
     zoobc.Network.list(hosts);
-    await expect(hosts).to.be.an('array');
+    expect(hosts).to.be.an('array');
   });
 
-  it('Network selected should return object', async () => {
+  it('should return object from selected', async () => {
     zoobc.Network.set(0);
     const resp = await zoobc.Network.selected();
-    await expect(resp).to.be.an('object');
+    expect(resp).to.be.an('object');
   });
 
-  it('Set network should return value 172.104.47.168:7000', async () => {
+  it('should return host value is 172.104.47.168:7000', async () => {
     zoobc.Network.set(1);
     const resp = await zoobc.Network.selected();
-    await expect(resp.host).to.equal('172.104.47.168:7000');
+    expect(resp.host).to.equal(hosts[1].host);
+  });
+
+  it('should return string from health check network', done => {
+    zoobc.Network.set(1);
+    zoobc.Network.ping().then(resp => {
+      expect(resp).to.be.an('string');
+    });
+    done();
   });
 });