+++ types/Transactions.d.ts
@@ -1,4 +1,4 @@
-import { PostTransactionResponse, GetTransactionsResponse, Transaction } from '../grpc/model/transaction_pb';
+import { PostTransactionResponse, GetTransactionsResponse, Transaction, GetTransactionMinimumFeeResponse } from '../grpc/model/transaction_pb';
 import { SendMoneyInterface } from './helper/transaction-builder/send-money';
 import { BIP32Interface } from 'bip32';
 export interface TransactionListParams {
@@ -16,9 +16,11 @@ export interface TransactionListParams {
 declare function getList(params?: TransactionListParams): Promise<GetTransactionsResponse.AsObject>;
 declare function get(id: string): Promise<Transaction.AsObject>;
 declare function sendMoney(data: SendMoneyInterface, seed: BIP32Interface): Promise<PostTransactionResponse.AsObject>;
+declare function getTransactionMinimumFee(data: SendMoneyInterface, seed: BIP32Interface): Promise<GetTransactionMinimumFeeResponse.AsObject>;
 declare const _default: {
     sendMoney: typeof sendMoney;
     get: typeof get;
     getList: typeof getList;
+    getTransactionMinimumFee: typeof getTransactionMinimumFee;
 };
 export default _default;